tests/wpt/tests/soft-navigation-heuristics/largest-painted-element-paint-first.tentative.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>SoftNavigationEntry getLargestInteractionContentfulPaint() method: Paint first, then URL.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/soft-navigation-helper.js"></script>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<main id=main>
  <a id=link>Click me!</a>
</main>
<script>
  promise_test(async t => {
    const soft_nav_promise = getNextEntry('soft-navigation');
    const clickTarget = document.getElementById("link");
    let largeId;

    clickTarget.addEventListener('click', async () => {
      // Yield to ensure task tracking is working.
      await new Promise(r => t.step_timeout(r, 0));

      // Paint first.
      const large = await addTextParagraphToMain("Much larger element that should be the largest interaction contentful paint in this interaction context");
      largeId = large.id;

      // URL change.
      history.pushState({}, '', `foobar.html?${counter++}`);
    });

    await test_driver.click(clickTarget);
    const entry = await soft_nav_promise;

    assert_true("getLargestInteractionContentfulPaint" in entry, "getLargestInteractionContentfulPaint method should exist on soft nav entry");
    assert_true("paintTime" in entry, "paintTime attribute should exist on soft nav entry");
    assert_greater_than(entry.paintTime, 0, "soft nav entry paintTime should be greater than 0");

    const largest = entry.getLargestInteractionContentfulPaint();
    assert_not_equals(largest, null, "getLargestInteractionContentfulPaint() is not null");
    assert_equals(largest.largestContentfulPaint.id, largeId, "getLargestInteractionContentfulPaint() has the correct ID");
    assert_true("paintTime" in largest.largestContentfulPaint, "paintTime attribute should exist on ICP entry");
    assert_greater_than(largest.largestContentfulPaint.paintTime, 0, "ICP entry paintTime should be greater than 0");
    assert_equals(largest.largestContentfulPaint.paintTime, entry.paintTime, "paintTime should match between entries");
  }, "Test that the getLargestInteractionContentfulPaint() method is correctly set when paint happens before URL change");
</script>
